E1: Finished exercise 1!
Overall, the background is constantly changing colour due to the mouseX & Y function, the house (square) grows in size at the beginning and is constrained when it reaches the center, the two aliens along with their eyes (Perry and Lola) move in opposite directions and reach at the same time(the square/house as well). The end!